AI_project/Project_work/requirements.txt
# ── Core Framework ─────────────────────────────────────────────────────────────
streamlit>=1.32.0

# ── LLM & AI ───────────────────────────────────────────────────────────────────
groq>=0.5.0

# ── Embeddings & Vector Search ─────────────────────────────────────────────────
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0

# ── Graph Database ─────────────────────────────────────────────────────────────
neo4j>=5.18.0

# ── NLP ────────────────────────────────────────────────────────────────────────
spacy>=3.7.0
# After install, run: python -m spacy download en_core_web_sm

# ── Numerical & ML ─────────────────────────────────────────────────────────────
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0

# ── PDF & Document Parsing ─────────────────────────────────────────────────────
pdfplumber>=0.10.0
python-docx>=1.1.0
reportlab>=4.1.0

# ── Environment Variables ──────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Standard Library (no install needed, listed for reference) ─────────────────
# os, sys, re, json, time, uuid, sqlite3, collections
